Rename src/print_string.h to PrintString.h, url_encoding.h to UrlEncoding.h
