+++ src/url_encoding/url_encoding.cpp
@@ -23,11 +23,11 @@ SOFTWARE.
 */
 
 #include <Print.h>
-#include "url_encoding.h"
+#include <UrlEncoding.h>
 
 // Inspired by
 // https://github.com/zenmanenergy/ESP8266-Arduino-Examples/blob/master/helloWorld_urlencoded/urlencode.ino.
-// Rewritten to use MemPrint instead of String.
+// Rewritten to use the PrintString object instead of String.
 
 namespace url_encoding {
