+++ tests/UrlEncodingTest/UrlEncodingTest.ino
@@ -1,8 +1,8 @@
 #line 2 "UrlEncodingTest.ino"
 
 #include <AUnit.h>
-#include <print_string.h>
-#include <url_encoding.h>
+#include <PrintString.h>
+#include <UrlEncoding.h>
 
 using namespace aunit;
 using namespace print_string;
@@ -13,16 +13,16 @@ using namespace url_encoding;
 //----------------------------------------------------------------------------
 
 test(formUrlEncode, normal) {
-  PrintString<10> memPrint;
-  formUrlEncode(memPrint, "0aA %");
-  assertEqual("0aA+%25", memPrint.getCstr());
+  PrintString<10> printString;
+  formUrlEncode(printString, "0aA %");
+  assertEqual("0aA+%25", printString.getCstr());
 }
 
 test(formUrlEncode, tooLong) {
-  PrintString<10> memPrint;
+  PrintString<10> printString;
   // truncated to 9 characters because PrintString is only 10-characters big
-  formUrlEncode(memPrint, "0aA %0123456");
-  assertEqual("0aA+%2501", memPrint.getCstr());
+  formUrlEncode(printString, "0aA %0123456");
+  assertEqual("0aA+%2501", printString.getCstr());
 }
 
 //----------------------------------------------------------------------------
@@ -30,30 +30,30 @@ test(formUrlEncode, tooLong) {
 //----------------------------------------------------------------------------
 
 test(formUrlDecode, normal) {
-  PrintString<10> memPrint;
-  formUrlDecode(memPrint, "0aA+%25");
-  assertEqual("0aA %", memPrint.getCstr());
+  PrintString<10> printString;
+  formUrlDecode(printString, "0aA+%25");
+  assertEqual("0aA %", printString.getCstr());
 }
 
 test(formUrlDecode, truncated1) {
-  PrintString<10> memPrint;
+  PrintString<10> printString;
   // Missing 2-digit of hex after %, which is ignored.
-  formUrlDecode(memPrint, "0aA+%2");
-  assertEqual("0aA ", memPrint.getCstr());
+  formUrlDecode(printString, "0aA+%2");
+  assertEqual("0aA ", printString.getCstr());
 }
 
 test(formUrlDecode, truncated2) {
-  PrintString<10> memPrint;
+  PrintString<10> printString;
   // Missing 2-digit of hex after %, which is ignored.
-  formUrlDecode(memPrint, "0aA+%");
-  assertEqual("0aA ", memPrint.getCstr());
+  formUrlDecode(printString, "0aA+%");
+  assertEqual("0aA ", printString.getCstr());
 }
 
 test(formUrlDecode, invalidHex) {
-  PrintString<10> memPrint;
+  PrintString<10> printString;
   // Missing 2-digit of hex after %, which is ignored.
-  formUrlDecode(memPrint, "0aA+%0Z");
-  assertEqual("0aA ", memPrint.getCstr());
+  formUrlDecode(printString, "0aA+%0Z");
+  assertEqual("0aA ", printString.getCstr());
 }
 
 //----------------------------------------------------------------------------
